pyproject.toml
[project]
name = "pdf-rag-qa"
version = "0.1.0"
description = ""
authors = [
    {name = "hari-testpress",email = "hari@testpress.in"}
]
readme = "README.md"


[tool.poetry.dependencies]
python = ">=3.11,<4.0"
python-dotenv = "^1.1.0"
marker-pdf = "^1.7.1"
ipython = "^9.2.0"
chonkie = {extras = ["semantic"], version = "^1.0.7"}
pydantic = "^2.11.4"
langchain-community = "^0.3.24"
langchain-openai = "^0.3.17"
faiss-cpu = "^1.11.0"
langchain = "^0.3.25"

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
